# Swagger_Documentation.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Employee Management API",
    "description": "APIs for managing employees",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/departments": {
      "get": {
        "tags": [
          "Department"
        ],
        "summary": "Get all departments",
        "description": "Retrieves a complete list of all departments within the organization.",
        "operationId": "getDepartments",
        "responses": {
          "200": {
            "description": "List of departments successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Department"
        ],
        "summary": "Create a new department",
        "description": "Creates a new department with the provided details such as name and description.",
        "operationId": "createDepartment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Department created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid department data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/attendances/checkout": {
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Record employee check-out",
        "description": "Records the check-out time for an employee when they leave work, marking the end of their attendance for the day.",
        "operationId": "checkOut",
        "parameters": [
          {
            "name": "attendanceId",
            "in": "query",
            "description": "Attendance record ID for check-out",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "checkOutTime",
            "in": "query",
            "description": "Employee's check-out time",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Check-out successfully recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Attendance record not found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/attendances/checkin": {
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Record employee check-in",
        "description": "Records the check-in time for an employee when they arrive at work, logging their attendance.",
        "operationId": "checkIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttendanceDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Check-in successfully recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid attendance data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/departments/{id}/employees": {
      "get": {
        "tags": [
          "Department"
        ],
        "summary": "Get all employees in a department",
        "description": "Retrieves a list of employees belonging to the specified department based on the department's ID.",
        "operationId": "getEmployeesByDepartmentId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the department to retrieve employees for",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of employees successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "404": {
            "description": "Department with the specified ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/attendances/{employeeId}": {
      "get": {
        "tags": [
          "Attendance"
        ],
        "summary": "Retrieve employee attendance records",
        "description": "Fetches all attendance records for a specific employee by their ID, including check-in and check-out times.",
        "operationId": "getAttendanceRecordsByEmployeeId",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "description": "Employee ID to retrieve attendance records",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attendance records successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceDTO"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/departments/{id}": {
      "delete": {
        "tags": [
          "Department"
        ],
        "summary": "Delete a department",
        "description": "Deletes the department identified by the given ID from the system.",
        "operationId": "deleteDepartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the department to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Department deleted successfully"
          },
          "404": {
            "description": "Department not found for the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "EmployeeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "departmentName": {
            "type": "string"
          },
          "joinDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "DepartmentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateDepartmentDTO": {
        "type": "object",
        "description": "The details of the department to be created",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "AttendanceDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "employeeName": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "checkIn": {
            "type": "string",
            "format": "time",
            "example": "01:01:01"
          },
          "checkOut": {
            "type": "string",
            "format": "time",
            "example": "01:01:01"
          }
        }
      },
      "CreateAttendanceDTO": {
        "type": "object",
        "description": "Attendance details for employee check-in",
        "properties": {
          "employeeId": {
            "type": "integer",
            "format": "int64"
          },
          "checkInTime": {
            "type": "string",
            "format": "time",
            "example": "01:01:01"
          }
        },
        "required": [
          "checkInTime",
          "employeeId"
        ]
      }
    }
  }
}